Save response value instead of printing temporary notices
If you go to the settings screen and you don't see an error, it makes it seem like everything is OK. When in fact, all that means is that the form has not been submitted.

I wanted the arrow to be persistent instead of temporary, so I saved the latest response.

This isn't perfect (if someone updates information on the TL admin it won't be cleared automatically), but it's better than not having any status indicators.